requirements.txt
# Core Data Processing
pandas==2.0.3
numpy==1.24.3

# PDF Processing
PyMuPDF==1.23.8
pdfplumber==0.10.3
camelot-py==0.11.0

# OCR
pytesseract==0.3.10

# HTML Processing
beautifulsoup4==4.12.2
lxml==4.9.3
trafilatura==1.6.3

# NLP & Language Detection
spacy==3.7.2
langdetect==1.0.9
nltk==3.8.1

# Text/Code Parsing
jsonschema==4.19.0

# Transformers (for embeddings, LM perplexity, NO KENLM)
transformers==4.34.1
torch==2.0.1
sentence-transformers==2.2.2

# Vector Search
faiss-cpu==1.7.4

# Deduplication
datasketch==1.0.8

# Web Scraping (optional)
requests==2.31.0

# UI & Web Frameworks
gradio==4.15.0
flask==3.0.0
streamlit==1.28.1

# Logging & Config
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
pydantic==2.4.2

# Task Scheduling (optional, for future)
APScheduler==3.10.4

# Code Parsing (optional)
ast-comment-parser==1.0.5
